access management customization

+++ app/Http/Controllers/SupervisorController.php
@@ -12,17 +12,25 @@ class SupervisorController extends Controller
 {
     public function index()
     {
-        $supervisors = User::where('role', 'supervisor')
-            ->where('dept_id', auth()->user()->dept_id)
-            ->orderBy('created_at', 'desc')
-            ->get();
-        return view('template.home.users.supervisors.index', compact('supervisors'));
+        if(auth()->user()->isSuperAdmin || auth()->user()->isAdmin){
+            $supervisors = User::where('role', 'supervisor')
+                ->where('dept_id', auth()->user()->dept_id)
+                ->orderBy('created_at', 'desc')
+                ->get();
+            return view('template.home.users.supervisors.index', compact('supervisors'));
+        } else {
+            return redirect('/');
+        }
     }
 
     public function create()
     {
-        $departments = Department::all();
-        return view('template.home.users.supervisors.create', compact('departments'));
+        if(auth()->user()->isSuperAdmin || auth()->user()->isAdmin){
+            $departments = Department::all();
+            return view('template.home.users.supervisors.create', compact('departments'));
+        } else {
+            return redirect('/');
+        }
     }
 
     public function createSupervisor()
@@ -33,28 +41,32 @@ class SupervisorController extends Controller
 
     public function store(Request $request)
     {
-        $request->validate([
-            'name' => ['required', 'string', 'max:255'],
-            'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:' . User::class],
-            'password' => ['required', 'confirmed', Rules\Password::defaults()],
-            // 'role' => 'required|string',
-            // 'isAdmin' => 'boolean',
-            // 'isSuperAdmin' => 'boolean',
-            'dept_id' => 'nullable|exists:departments,id',
-        ]);
-
-        $user = User::create([
-            'name' => $request->name,
-            'official_id' => $request->teacher_id,
-            'email' => $request->email,
-            'password' => Hash::make($request->password),
-            'designation' => $request->designation,
-            'role' => 'supervisor',
-            'isAdmin' => $request->isAdmin ?? false,
-            'isSuperAdmin' => $request->isSuperAdmin ?? false,
-            'dept_id' => auth()->user()->dept_id,
-        ]);
-        return redirect()->route('supervisors.index');
+        if(auth()->user()->isSuperAdmin || auth()->user()->isAdmin){
+            $request->validate([
+                'name' => ['required', 'string', 'max:255'],
+                'email' => ['required', 'string', 'lowercase', 'email', 'max:255', 'unique:' . User::class],
+                'password' => ['required', 'confirmed', Rules\Password::defaults()],
+                // 'role' => 'required|string',
+                // 'isAdmin' => 'boolean',
+                // 'isSuperAdmin' => 'boolean',
+                'dept_id' => 'nullable|exists:departments,id',
+            ]);
+    
+            $user = User::create([
+                'name' => $request->name,
+                'official_id' => $request->teacher_id,
+                'email' => $request->email,
+                'password' => Hash::make($request->password),
+                'designation' => $request->designation,
+                'role' => 'supervisor',
+                'isAdmin' => $request->isAdmin ?? false,
+                'isSuperAdmin' => $request->isSuperAdmin ?? false,
+                'dept_id' => auth()->user()->dept_id,
+            ]);
+            return redirect()->route('supervisors.index');
+        } else {
+            return redirect('/');
+        }
     }
 
     public function storeSupervisors(Request $request)
@@ -85,33 +97,44 @@ class SupervisorController extends Controller
 
     public function edit($id)
     {
-        $supervisor = User::findOrFail($id);
-        $departments = Department::all();
-        return view('template.home.users.supervisors.edit', compact('supervisor', 'departments'));
+        if(auth()->user()->isSuperAdmin || auth()->user()->isAdmin){
+            $supervisor = User::findOrFail($id);
+            $departments = Department::all();
+            return view('template.home.users.supervisors.edit', compact('supervisor', 'departments'));
+        } else {
+            return redirect('/');
+        }
     }
 
     public function update(Request $request, $id)
     {
-
-        $supervisor = User::findOrFail($id);
-        
-        $supervisor->update([
-            'name' => $request->name,
-            'email' => $request->email,
-            'designation' => $request->designation,
-            'role' => 'supervisor',
-            'isAdmin' => $request->isAdmin ?? false,
-            'isSuperAdmin' => $request->isSuperAdmin ?? false,
-        ]);
-
-        return redirect()->back();
+        if(auth()->user()->isSuperAdmin || auth()->user()->isAdmin){
+            $supervisor = User::findOrFail($id);
+            
+            $supervisor->update([
+                'name' => $request->name,
+                'email' => $request->email,
+                'designation' => $request->designation,
+                'role' => 'supervisor',
+                'isAdmin' => $request->isAdmin ?? false,
+                'isSuperAdmin' => $request->isSuperAdmin ?? false,
+            ]);
+    
+            return redirect()->back();
+        } else {
+            return redirect('/');
+        }
     }
 
     public function destroy($id)
     {
-        $supervisor = User::findOrFail($id);
-        $supervisor->delete();
-
-        return redirect()->back();
+        if(auth()->user()->isSuperAdmin || auth()->user()->isAdmin){
+            $supervisor = User::findOrFail($id);
+            $supervisor->delete();
+    
+            return redirect()->back();
+        } else {
+            return redirect('/');
+        }
     }
 }
